works on mac

+++ include/tos/span.hpp
@@ -14,12 +14,12 @@ namespace tos
     public:
         using iterator = T*;
 
-        constexpr span(T* base, size_t len) : m_base(base), m_len(len) {}
+        constexpr span(T* base, size_t len) : m_base(base), m_len(ptrdiff_t(len)) {}
 
         constexpr span(T* base, T* end) : m_base{base}, m_len{end - base} {}
 
         template <size_t Sz>
-        constexpr span(T (&arr)[Sz]) : m_base(arr), m_len(Sz) {}
+        constexpr span(T (&arr)[Sz]) : m_base(arr), m_len(ptrdiff_t(Sz)) {}
 
         constexpr size_t size() const { return m_len; }
 
@@ -41,7 +41,7 @@ namespace tos
         constexpr const T* end() const { return m_base + m_len; }
 
         constexpr operator span<const T>() const {
-            return { m_base, m_len };
+            return { m_base, size_t(m_len) };
         }
 
         constexpr span slice(size_t begin, size_t len){
